+++ src/extractors/isbnsearch.js
@@ -1,0 +1,3 @@
+function getIsbnSearchDetails() {
+    
+}
