+++ src/extractors/storygraph.js
@@ -1,0 +1,5 @@
+function getStoryGraphDetails() {
+    console.log('[👩🏻‍🏫 Marian] Extracting The StoryGraph details');
+}
+
+export { getStoryGraphDetails };
